+++ include/vsmc/core/core.hpp
@@ -35,7 +35,6 @@
 #include <vsmc/internal/config.h>
 #include <vsmc/core/monitor.hpp>
 #include <vsmc/core/particle.hpp>
-#include <vsmc/core/path.hpp>
 #include <vsmc/core/sampler.hpp>
 #include <vsmc/core/single_particle.hpp>
 #include <vsmc/core/state_matrix.hpp>