+++ include/vsmc/smp/backend_tbb.hpp
@@ -32,8 +32,8 @@
 #ifndef VSMC_SMP_BACKEND_TBB_HPP
 #define VSMC_SMP_BACKEND_TBB_HPP
 
-#include <tbb/tbb.h>
 #include <vsmc/smp/backend_base.hpp>
+#include <tbb/tbb.h>
 
 #define VSMC_DEFINE_SMP_BACKEND_TBB_PARALLEL_RUN_INITIALIZE(args)             \
     this->eval_param(particle, param);                                        \
@@ -55,13 +55,6 @@
     work_type work(this, iter, dim, &particle, r);                            \
     ::tbb::parallel_for args;                                                 \
     this->eval_post(iter, particle);
-
-#define VSMC_DEFINE_SMP_BACKEND_TBB_PARALLEL_RUN_PATH_EVAL(args)              \
-    this->eval_pre(iter, particle);                                           \
-    work_type work(this, iter, &particle, r);                                 \
-    ::tbb::parallel_for args;                                                 \
-    this->eval_post(iter, particle);                                          \
-    return this->eval_grid(iter, particle);
 
 namespace vsmc
 {
@@ -502,106 +495,6 @@ class MonitorEvalTBB : public MonitorEvalBase<T, Derived>
     }
 #endif // __TBB_TASK_GROUP_CONTEXT
 };     // class MonitorEvalTBB
-
-/// \brief Path<T>::eval_type subtype using Intel Threading Building Blocks
-/// \ingroup TBB
-template <typename T, typename Derived>
-class PathEvalTBB : public PathEvalBase<T, Derived>
-{
-    public:
-    double operator()(std::size_t iter, Particle<T> &particle, double *r)
-    {
-        return parallel_run(iter, particle, r,
-            ::tbb::blocked_range<typename Particle<T>::size_type>(
-                                0, particle.size()));
-    }
-
-    protected:
-    VSMC_DEFINE_SMP_BACKEND_SPECIAL(TBB, PathEval)
-
-    class work_type
-    {
-        public:
-        using size_type = typename Particle<T>::size_type;
-
-        work_type(PathEvalTBB<T, Derived> *wptr, std::size_t iter,
-            Particle<T> *pptr, double *r)
-            : wptr_(wptr), iter_(iter), pptr_(pptr), r_(r)
-        {
-        }
-
-        void operator()(const ::tbb::blocked_range<size_type> &range) const
-        {
-            for (size_type i = range.begin(); i != range.end(); ++i)
-                r_[i] = wptr_->eval_sp(iter_, pptr_->sp(i));
-        }
-
-        private:
-        PathEvalTBB<T, Derived> *const wptr_;
-        const std::size_t iter_;
-        Particle<T> *const pptr_;
-        double *const r_;
-    }; // class ParallelPathState
-
-    double parallel_run(std::size_t iter, Particle<T> &particle, double *r,
-        const ::tbb::blocked_range<typename Particle<T>::size_type> &range)
-    {
-        VSMC_DEFINE_SMP_BACKEND_TBB_PARALLEL_RUN_PATH_EVAL((range, work));
-    }
-
-    double parallel_run(std::size_t iter, Particle<T> &particle, double *r,
-        const ::tbb::blocked_range<typename Particle<T>::size_type> &range,
-        const ::tbb::auto_partitioner &partitioner)
-    {
-        VSMC_DEFINE_SMP_BACKEND_TBB_PARALLEL_RUN_PATH_EVAL(
-            (range, work, partitioner));
-    }
-
-    double parallel_run(std::size_t iter, Particle<T> &particle, double *r,
-        const ::tbb::blocked_range<typename Particle<T>::size_type> &range,
-        const ::tbb::simple_partitioner &partitioner)
-    {
-        VSMC_DEFINE_SMP_BACKEND_TBB_PARALLEL_RUN_PATH_EVAL(
-            (range, work, partitioner));
-    }
-
-    double parallel_run(std::size_t iter, Particle<T> &particle, double *r,
-        const ::tbb::blocked_range<typename Particle<T>::size_type> &range,
-        ::tbb::affinity_partitioner &partitioner)
-    {
-        VSMC_DEFINE_SMP_BACKEND_TBB_PARALLEL_RUN_PATH_EVAL(
-            (range, work, partitioner));
-    }
-
-#if __TBB_TASK_GROUP_CONTEXT
-    double parallel_run(std::size_t iter, Particle<T> &particle, double *r,
-        const ::tbb::blocked_range<typename Particle<T>::size_type> &range,
-        const ::tbb::auto_partitioner &partitioner,
-        ::tbb::task_group_context &context)
-    {
-        VSMC_DEFINE_SMP_BACKEND_TBB_PARALLEL_RUN_PATH_EVAL(
-            (range, work, partitioner, context));
-    }
-
-    double parallel_run(std::size_t iter, Particle<T> &particle, double *r,
-        const ::tbb::blocked_range<typename Particle<T>::size_type> &range,
-        const ::tbb::simple_partitioner &partitioner,
-        ::tbb::task_group_context &context)
-    {
-        VSMC_DEFINE_SMP_BACKEND_TBB_PARALLEL_RUN_PATH_EVAL(
-            (range, work, partitioner, context));
-    }
-
-    double parallel_run(std::size_t iter, Particle<T> &particle, double *r,
-        const ::tbb::blocked_range<typename Particle<T>::size_type> &range,
-        ::tbb::affinity_partitioner &partitioner,
-        ::tbb::task_group_context &context)
-    {
-        VSMC_DEFINE_SMP_BACKEND_TBB_PARALLEL_RUN_PATH_EVAL(
-            (range, work, partitioner, context));
-    }
-#endif // __TBB_TASK_GROUP_CONTEXT
-};     // PathEvalTBB
 
 } // namespace vsmc